fix(history): isolate Markdown-aware display dependencies
