+++ src/auto-reply/reply/display-text-sanitize.ts
@@ -1,14 +1,9 @@
 // Removes internal runtime context from text shown back to users.
+// Keep this leaf free of Markdown parsing: terminal and transformed SDK paths import it.
 import { stripInternalRuntimeContext } from "../../agents/internal-runtime-context.js";
-import { stripEnvelope, stripMessageIdHints } from "../../shared/chat-envelope.js";
 import { stripInboundMetadata } from "./strip-inbound-meta.js";
 
 /** Removes internal runtime metadata before showing text to users. */
 export function stripInternalMetadataForDisplay(text: string): string {
   return stripInboundMetadata(stripInternalRuntimeContext(text));
-}
-
-/** Removes user-envelope and message-id hints from display text. */
-export function stripUserEnvelopeForDisplay(text: string): string {
-  return stripMessageIdHints(stripEnvelope(stripInternalMetadataForDisplay(text)));
 }
